fix: fixed count data type  parameter in handle_type function to expected int

+++ ft_printf.c
@@ -12,7 +12,7 @@
 
 #include "ft_printf.h"
 
-static int	handle_type(const char type, va_list args, size_t *count)
+static int	handle_type(const char type, va_list args, int *count)
 {
 	if (type == 'd' || type == 'i')
 		pf_putnbr(va_arg(args, int), count);
